환불 아임포트

+++ adidas_store_tn/src/main/java/kr/green/adidas/service/OrderServiceImp.java
@@ -116,6 +116,8 @@ public class OrderServiceImp implements OrderService{
 			return;
 		if(orderList.getOl_num() <= 0)
 			return;
+		if(orderList.getOl_or_num() <= 0)
+			return;
 		orderDao.deleteOrderList(orderList);
 	}
